Move string tests to appropriate file
