+++ dig_test.go
@@ -723,22 +723,3 @@ func TestInvokeFailures(t *testing.T) {
 		assert.Error(t, c.Invoke(func() (int, error) { return 42, errors.New("oh no") }))
 	})
 }
-
-func TestStringer(t *testing.T) {
-	c := New()
-	type A struct{}
-	type B struct{}
-	require.NoError(t, c.Provide(func() (*A, *B) { return &A{}, &B{} }))
-	require.NoError(t, c.Invoke(func(a *A) {}))
-
-	b := &bytes.Buffer{}
-	fmt.Fprintln(b, c)
-	s := b.String()
-
-	assert.Contains(t, s, "*dig.A ->")
-	assert.Contains(t, s, "*dig.B ->")
-	assert.Contains(t, s, "func() (*dig.A, *dig.B)")
-
-	assert.Contains(t, s, "*dig.A =>")
-	assert.Contains(t, s, "*dig.B =>")
-}
